add a simple heap problem 😭
I've learnt in this problem that MinGW implementations of scanf and
printf are very slow. To solve it we should define a macro to tell MinGW
not to use those implementations.

  #define __USE_MINGW_ANSI_STDIO 0

I've gone through hell to find this out. 😢

It's worth noting that this issues doesn not exist in g++ compiler
version. So if you submit you c code as a c++ code the execution time
increases.

These are some links to investigate it more deeply:

https://codeforces.com/blog/entry/68043
https://codeforces.com/blog/entry/105687
https://codeforces.com/blog/entry/95872
https://codeforces.com/blog/entry/47180